pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "getlegit"
version = "0.1.0"
description = "The trust layer for AI agents"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
authors = [
    { name = "Legit Contributors" },
]
keywords = ["ai", "agents", "benchmark", "trust", "evaluation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "jsonschema>=4.20",
]

[project.scripts]
legit = "getlegit.cli.main:cli"

[project.urls]
Homepage = "https://getlegit.dev"
Repository = "https://github.com/getlegitdev/legit"
Documentation = "https://github.com/getlegitdev/legit/tree/main/docs"
Issues = "https://github.com/getlegitdev/legit/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/getlegit"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
